# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Dead Sea Levant Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/home": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Returns Home data",
                "description": "Returns Home data",
                "operationId": "Home",
                "responses": {
                    "200": {
                        "description": "Returns Home data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/HomeResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong input response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "properties": {
                                                "dynamic-error-keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contact_requests": {
            "post": {
                "tags": [
                    "Home"
                ],
                "summary": "Send A Contact Request",
                "description": "Send A Contact Request",
                "operationId": "ContactRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactRequestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "success"
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "properties": {
                                                "dynamic-error-keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "List Products",
                "description": "List Products",
                "operationId": "ListProducts",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "category id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "availablity",
                        "in": "query",
                        "description": "availablity",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "price_from",
                        "in": "query",
                        "description": "price from",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price_to",
                        "in": "query",
                        "description": "price to",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of Products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string"
                                                },
                                                "next": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string"
                                                            },
                                                            "label": {
                                                                "type": "string"
                                                            },
                                                            "active": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong input response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "properties": {
                                                "dynamic-error-keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ContactRequestRequest": {
                "title": "Contact Request Request",
                "description": "Request body for Contact Request Request",
                "required": [
                    "name",
                    "email",
                    "subject",
                    "message"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "subject": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CategoryResource": {
                "title": "Category Resource",
                "description": "Resource of the Category object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "HomeResource": {
                "title": "Home Resource",
                "description": "Resource of the Home object",
                "properties": {
                    "translations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LangFileResource"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ImageResource"
                        }
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryResource"
                        }
                    }
                },
                "type": "object"
            },
            "ImageResource": {
                "title": "Image Resource",
                "description": "Resource of the Image object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LangFileResource": {
                "title": "Lang File Resource",
                "description": "Resource of the Lang File object",
                "properties": {
                    "lang_file_name": {
                        "type": "object",
                        "additionalProperties": [
                            {
                                "type": "string",
                                "example": "value",
                                "property": "key"
                            },
                            {
                                "type": "string",
                                "example": "value2",
                                "property": "key2"
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "ProductImagesResource": {
                "title": "Product Images Resource",
                "description": "Resource of the Product Images object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "title": "Product Resource",
                "description": "Resource of the Product object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryResource"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "stock": {
                        "type": "number",
                        "format": "double"
                    },
                    "shipping_terms": {
                        "type": "string"
                    },
                    "product_reviews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductReviewsResource"
                        }
                    },
                    "product_images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImagesResource"
                        }
                    }
                },
                "type": "object"
            },
            "ProductReviewsResource": {
                "title": "ProductReviewsResource",
                "description": "Resource of the Product Reviews object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "stars": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "customer_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}